fix sparge water temp calculation
Fix sparge water temperature calculation
Calculation was correctly using gallons absorbed by grain,
but needed to multiply by the weight of water to get a mass!

also changing grain_weight to dry_grain_weight for clarity
and adding a docstring to calculate_water_absorbed as a reminder
that it returns gallons absorbed, not mass absorbed.